Modularize negotiation buffer utilities
