+++ tools/line_limits.toml
@@ -53,11 +53,6 @@ max_lines = 6400
 warn_lines = 6300
 
 [[overrides]]
-path = "crates/transport/src/negotiation/buffer.rs"
-max_lines = 1200
-warn_lines = 1100
-
-[[overrides]]
 path = "crates/transport/src/negotiation/parts.rs"
 max_lines = 1100
 warn_lines = 1000
